Add improvements to walls and collision detecting

+++ app/src/main/java/com/example/dungeoncrawler/model/Chest.java
@@ -13,7 +13,7 @@ public class Chest {
     public static Location getLocation() {
         return location;
     }
-    public static void setLocation(float newX, float newY) {
+    public static void setLocation(int newX, int newY) {
         Chest.location.setxCord(newX);
         Chest.location.setyCord(newY);
     }
